I have added the Ratskin house along with weapons
for them and the scavvies. Some of the weapons are not finished yet. I did
some refactoring to make the redemptionist class file names consistent
with the file names of the other outlander houses and to keep the files in
order. I made the table larger for bigger scale fights and had to change
the NeoTexture files to look sharper on the bigger table. I then started playing
around with the NeoTexture files (by code, without using the NeoTexture
plugin because I couldn't find it) for different looking textures. I have
been tinkering with new skyboxes for a while and after modifying the toxic
atmosphere box which I found on a free texture site, I decided it didn't
look right. I then made my own skybox with a hd panorama shot of tokyo I
found on google :) It makes the ceiling (or the wall as its known as in
the fluff) to look more like there is a vast city above and gives more sense
of scale to the underhive. From the scaled layout of hive primus in the latest
rulebook, it is about 1 mile from the wall to the uninhabitable sump.

I made some changes to the lighting to make it darker. I was also trying to
make the miniature images darker but still haven't worked out how to do
that yet. I moved the Bounty Hunter advance rolls to their profile and the
scavvy followers code to the scavvy leaders profile. The code for the
advance rolls and the followers is running twice and I don't know why. So
the bounty hunter will get 2 random points (and sometimes alot of points
for weapon skill) instead of 1 and the scavvies will get two types of
random followers. That needs fixing. I don't think I changed anything
else. I need to start making change lists :)